requirements.txt
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
diffusers==0.14.0
filelock==3.9.0
flake8==5.0.4
huggingface-hub==0.13.1
idna==3.4
importlib-metadata==6.0.0
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
Pillow==9.4.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==1.0.0
pytube==12.1.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.15
zipp==3.15.0
